+++ enum/src/main/java/org/grnet/status/enums/InvitationAction.java
@@ -2,5 +2,6 @@ package org.grnet.status.enums;
 
 public enum InvitationAction {
     ACCEPT,
-    REJECT
+    REJECT,
+    REVOKE
 }