+++ enum/src/main/java/org/grnet/status/enums/InvitationStatus.java
@@ -3,5 +3,6 @@ package org.grnet.status.enums;
 public enum InvitationStatus {
     PENDING,
     ACCEPTED,
-    REJECTED
+    REJECTED,
+    REVOKED
 }